calendar_agent/requirements.txt
google-auth>=2.16.2
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.86.0
openai>=1.0.0
python-dotenv>=1.0.0